+++ python/run_pipeline.py
@@ -56,7 +56,7 @@ def run_pipeline(inputdir, outputdir, inference_method=None, config_path=None):
 
     # Rule: If inference_method is not provided (None), use the one from config.json
     if inference_method is None:
-        inference_method = ANALYSIS_PARAMETERS.get("inference_method", "basil")  # Default to 'basil' if not in config
+        inference_method = ANALYSIS_PARAMETERS.get("inference_method", "vaby")  # Default to 'vaby' if not in config
     else:
         # Override the config's inference_method with the user-provided value
         ANALYSIS_PARAMETERS["inference_method"] = inference_method
@@ -92,7 +92,7 @@ def main():
         epilog="""
     Examples:
     python run_pipeline.py /input /output
-    python run_pipeline.py /input /output --inference-method basil
+    python run_pipeline.py /input /output --inference-method vaby for BASIL-like output
     python run_pipeline.py /input /output --inference-method ssvb --config /path/to/config.json
         """
     )
@@ -102,8 +102,8 @@ def main():
     parser.add_argument("outputdir", type=str, 
                         help="Path to output directory for DICOM results for PACS and IMAGER platform")
     parser.add_argument("--inference-method", type=str, default=None, 
-                        choices=["basil", "ssvb"],
-                        help="Optional input to choose inference method for fitting: 'basil' or 'ssvb'. If not provided, uses the value from config.json.")
+                        choices=["vaby", "ssvb"],
+                        help="Optional input to choose inference method for fitting: 'vaby' or 'ssvb'. If not provided, uses the value from config.json.")
     parser.add_argument("--config", type=str, default=None,
                         help="Optional path to config.json with processing parameters")
     parser.add_argument("--version", action="version", version=f"ClinicalASL {TOOL_VERSION}")